Implementacion del producto

+++ Infraestructure/Models/Entities/Medida.cs
@@ -4,9 +4,9 @@ namespace ab_accesorios_be.Infraestructure.Models.Entities
 {
     public class Medida : EntityBase
     {
-        public int Alto { get; set; }
-        public int Ancho { get; set; }
-        public int Profundidad { get; set; }
+        public float Alto { get; set; }
+        public float Ancho { get; set; }
+        public float Profundidad { get; set; }
         public long ProductoId { get; set; }
         public Producto? Producto { get; set; }
     }
